requirements.txt
openai==1.52.0
requests
Pillow
opencv-python
numpy
pandas
PyYAML